Making Authorized Requests

+++ frontend/src/components/WorkoutDetails.js
@@ -1,13 +1,21 @@
-import { useDispatch } from "react-redux";
+import { useSelector, useDispatch } from "react-redux";
 import { deleteWorkout } from "../rtk/slices/workout-slice";
 
 // date fns
 import formatDistanceToNow from "date-fns/formatDistanceToNow";
 const WorkoutDetails = ({ workout }) => {
+  const user = useSelector((state) => state.auth);
   const dispatch = useDispatch();
   const handleClick = async () => {
+    if (!user) {
+      return;
+    }
+
     const response = await fetch("/api/workouts/" + workout._id, {
       method: "DELETE",
+      headers: {
+        Authorization: `Bearer ${user.token}`,
+      },
     });
     const json = await response.json();
